+++ src/main/java/com/turkcell/rentACar/business/concretes/ColorManager.java
@@ -1,12 +1,18 @@
 package com.turkcell.rentACar.business.concretes;
 
 import com.turkcell.rentACar.business.abstracts.ColorService;
+import com.turkcell.rentACar.business.dtos.BrandListDto;
+import com.turkcell.rentACar.business.dtos.ColorDto;
 import com.turkcell.rentACar.business.dtos.ColorListDto;
-import com.turkcell.rentACar.business.requests.CreateColorRequest;
-import com.turkcell.rentACar.business.requests.DeleteColorRequest;
-import com.turkcell.rentACar.business.requests.UpdateColorRequest;
+import com.turkcell.rentACar.business.requests.creates.CreateColorRequest;
+import com.turkcell.rentACar.business.requests.deletes.DeleteColorRequest;
+import com.turkcell.rentACar.business.requests.updates.UpdateColorRequest;
+import com.turkcell.rentACar.core.utilities.exceptions.BusinessException;
 import com.turkcell.rentACar.core.utilities.mapping.ModelMapperManager;
+import com.turkcell.rentACar.core.utilities.results.*;
+import com.turkcell.rentACar.dataAccess.abstracts.CarDao;
 import com.turkcell.rentACar.dataAccess.abstracts.ColorDao;
+import com.turkcell.rentACar.entities.concretes.Car;
 import com.turkcell.rentACar.entities.concretes.Color;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -19,56 +25,70 @@ public class ColorManager implements ColorService {
 
     private ColorDao colorDao;
     private ModelMapperManager modelMapperService;
+    private CarDao carDao;
 
     @Autowired
-    public ColorManager(ColorDao colorDao,ModelMapperManager modelMapperService) {
+    public ColorManager(ColorDao colorDao,ModelMapperManager modelMapperService,CarDao carDao) {
 
         this.colorDao = colorDao;
         this.modelMapperService = modelMapperService;
+        this.carDao = carDao;
     }
 
     @Override
-    public List<ColorListDto> getAll() {
-        
+    public DataResult<List<ColorListDto>> getAll() {
         List<Color> result = this.colorDao.findAll();
+        if(result.isEmpty())
+            return new ErrorDataResult<List<ColorListDto>>(null,"Current List is Empty");
         List<ColorListDto> response = result.stream().map(color->this.modelMapperService.forDto().map(color,ColorListDto.class)).collect(Collectors.toList());
-        return response;
+        return new SuccessDataResult<List<ColorListDto>>(response,"Colors Listed Successfully");
     }
 
     @Override
-    public void add(CreateColorRequest createColorRequest) throws Exception {
+    public Result add(CreateColorRequest createColorRequest) throws BusinessException {
         Color color = this.modelMapperService.forRequest().map(createColorRequest,Color.class);
         checkIfSameColor(color);
         this.colorDao.save(color);
+        return new SuccessResult("Color Added Successfully");
     }
 
     @Override
-    public ColorListDto getById(int id) {
+    public DataResult<ColorDto> getById(int id) {
+        if(this.colorDao.getAllByColorId(id).stream().count()==0)
+            return new ErrorDataResult(null,"Color Does Not Exist");
         Color color = colorDao.getById(id);
-        ColorListDto colorListDto = this.modelMapperService.forDto().map(color,ColorListDto.class);
-        return colorListDto;
+        ColorDto colorDto = this.modelMapperService.forDto().map(color,ColorDto.class);
+        return new SuccessDataResult<ColorDto>(colorDto,"Color Listed Successfully");
     }
 
     @Override
-    public void update(UpdateColorRequest updateColorRequest) {
+    public Result update(UpdateColorRequest updateColorRequest) throws BusinessException{
+        if(this.colorDao.getAllByColorId(updateColorRequest.getColorId()).stream().count()==0)
+            return new ErrorResult("Brand Does Not Exist");
         Color color = colorDao.getById(updateColorRequest.getColorId());
+        checkIfSameColor(color);
         color = this.modelMapperService.forRequest().map(updateColorRequest,Color.class);
         this.colorDao.save(color);
+        return new SuccessResult("Color Updated Succesfully");
     }
 
     @Override
-    public void delete(DeleteColorRequest DeleteColorRequest) {
-        Color color = this.modelMapperService.forRequest().map(DeleteColorRequest,Color.class);
+    public Result delete(DeleteColorRequest deleteColorRequest) throws BusinessException{
+        if(this.colorDao.getAllByColorId(deleteColorRequest.getColorId()).stream().count()==0)
+            return new ErrorResult("Brand Does Not Exist");
+        Color color = this.modelMapperService.forRequest().map(deleteColorRequest,Color.class);
+        checkIfColorUsedOnCar(color);
         this.colorDao.delete(color);
+        return new SuccessResult("Color Deleted Succesfully");
     }
 
-    void checkIfSameColor(Color color) throws Exception {
+    void checkIfSameColor(Color color) throws BusinessException {
         if(this.colorDao.getAllByColorName(color.getColorName()).stream().count()!=0)
-            throw new Exception("This brand already exists");
-        /*List<Color> colors = colorDao.findAll();
-        for (Color c:colors) {
-            if(c.getName().equals(color.getName()))
-                throw new Exception("This color already exists");
-        }*/
+            throw new BusinessException("This brand already exists");
+    }
+
+    void checkIfColorUsedOnCar(Color color) throws BusinessException{
+        if(this.carDao.getAllByColor_ColorId(color.getColorId()).stream().count()!=0)
+            throw new BusinessException("This Color Used on Cars ,You Can not Delete This Color");
     }
 }
